Ticket CHECK-158: Refactoring and adding a GraphQL field / class method to return the current enabled integrations for the tipline and their information from Smooch API
