+++ app/graph/types/team_bot_installation_type.rb
@@ -23,6 +23,8 @@ TeamBotInstallationType = GraphqlCrudOperations.define_default_type do
   end
 
   # Only for Smooch Bot
+  
+  field :smooch_enabled_integrations, JsonStringType
 
   field :smooch_bot_preview_rss_feed do
     type types.String
