+++ app/models/team_bot_installation.rb
@@ -21,13 +21,6 @@ class TeamBotInstallation < TeamUser
     BotUser.where(id: self.user_id).last
   end
 
-  def reset_smooch_authorization_token
-    token = SecureRandom.hex
-    self.set_smooch_authorization_token = token
-    self.set_smooch_twitter_authorization_url = "#{CheckConfig.get('checkdesk_base_url')}/api/users/auth/twitter?context=smooch&destination=#{CheckConfig.get('checkdesk_base_url')}/api/admin/smooch_bot/#{self.id}/authorize/twitter?token=#{token}"
-    self.set_smooch_facebook_authorization_url = "#{CheckConfig.get('checkdesk_base_url')}/api/users/auth/facebook?context=smooch&destination=#{CheckConfig.get('checkdesk_base_url')}/api/admin/smooch_bot/#{self.id}/authorize/facebook?token=#{token}"
-  end
-
   def apply_default_settings
     bot = self.bot_user
     if !bot.blank? && !bot.get_settings.blank?
